HH-218702 move RoutingDataSourceFactory to nab-jdbc module
