+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/routing/RoutingDataSourceFactory.java
@@ -1,9 +1,8 @@
-package ru.hh.nab.hibernate.datasource;
+package ru.hh.nab.jdbc.routing;
 
 import jakarta.inject.Named;
 import javax.sql.DataSource;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.jdbc.routing.RoutingDataSource;
 import ru.hh.nab.metrics.StatsDSender;
 
 public class RoutingDataSourceFactory {
