+++ nab-tests/src/test/java/ru/hh/nab/jpa/DataSourceSwitchingTest.java
@@ -45,7 +45,6 @@ import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.healthcheck.HealthCheckResultImpl;
 import ru.hh.nab.datasource.healthcheck.UnhealthyDataSourceException;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
-import ru.hh.nab.hibernate.datasource.RoutingDataSourceFactory;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.DataSourceType;
 import ru.hh.nab.jdbc.common.DatabaseSwitcher;
@@ -53,6 +52,7 @@ import ru.hh.nab.jdbc.common.healthcheck.HealthCheckResult;
 import static ru.hh.nab.jdbc.routing.DataSourceContext.onDataSource;
 import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 import ru.hh.nab.jdbc.routing.RoutingDataSource;
+import ru.hh.nab.jdbc.routing.RoutingDataSourceFactory;
 import ru.hh.nab.jpa.model.TestEntity;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
